my hw

+++ LoanCalc.java
@@ -13,9 +13,9 @@ public class LoanCalc {
      */
     public static void main(String[] args) {
         // Gets the loan data
-        double loan = Integer.parseInt(args[0]);
-        double rate = Integer.parseInt(args[1]);
-        int n = Integer.parseInt(args[2]);
+        double loan = Double.parseDouble(args[0]);
+        double rate = Double.parseDouble(args[1]);
+        int n = Double.parseDouble(args[2]);
 
         System.out.println("Loan sum = " + loan + ", interest rate = " + rate + "%, periods = " + n);
